requirements.txt
# Consult release notes for supported versions of Django, Wagtail, and Python.
# https://docs.coderedcorp.com/wagtail-crx/releases/
coderedcms==3.0.3
wagtail==5.2.*

astropy==6.0.0
django_bootstrap_icons==0.8.7
django_extensions==3.2.3
django-qr-code==3.1.1
django-social-share==2.3.0
ephem==4.1.5
fontawesomefree==6.5.1
pylunar==0.6.0
pythaidate==0.2.0
skyfield==1.46
wand==0.6.13
wagtail-modeladmin==1.0.0
infobip-api-python-client==3.0.3
crispy-bootstrap5==2023.10
django-crispy-forms==2.1
django-htmx==1.17.2
pyotp==2.9.0
django-formtools==2.5.1
python-dotenv==1.0.1
